requirements.txt
python-telegram-bot
requests
python-dotenv
python-telegram-bot[job-queue]